Move query and mutation to a new fetch directory
