+++ packages/ra-core/src/fetch/index.ts
@@ -1,0 +1,19 @@
+import HttpError from './HttpError';
+import * as fetchUtils from './fetch';
+import Mutation from './Mutation';
+import Query from './Query';
+import useDataProvider from './useDataProvider';
+import useMutation from './useMutation';
+import useQuery from './useQuery';
+import withDataProvider from './withDataProvider';
+
+export {
+    fetchUtils,
+    HttpError,
+    Mutation,
+    Query,
+    useDataProvider,
+    useMutation,
+    useQuery,
+    withDataProvider,
+};
