+++ packages/ra-core/src/index.ts
@@ -22,6 +22,7 @@ export {
 export * from './dataFetchActions';
 export * from './actions';
 export * from './auth';
+export * from './fetch';
 export * from './i18n';
 export * from './inference';
 export * from './util';
