+++ packages/ra-core/src/util/index.ts
@@ -3,39 +3,23 @@ import FieldTitle from './FieldTitle';
 import getFetchedAt from './getFetchedAt';
 import getFieldLabelTranslationArgs from './getFieldLabelTranslationArgs';
 import ComponentPropType from './ComponentPropType';
-import HttpError from './HttpError';
 import linkToRecord from './linkToRecord';
-import Mutation from './Mutation';
-import Query from './Query';
 import removeEmpty from './removeEmpty';
 import removeKey from './removeKey';
 import resolveRedirectTo from './resolveRedirectTo';
 import TestContext from './TestContext';
-import useDataProvider from './useDataProvider';
-import useMutation from './useMutation';
-import useQuery from './useQuery';
 import warning from './warning';
-import withDataProvider from './withDataProvider';
-import * as fetchUtils from './fetch';
 
 export {
-    fetchUtils,
     downloadCSV,
     FieldTitle,
     getFetchedAt,
     getFieldLabelTranslationArgs,
     ComponentPropType,
-    HttpError,
     linkToRecord,
-    Mutation,
-    Query,
     removeEmpty,
     removeKey,
     resolveRedirectTo,
     TestContext,
-    useDataProvider,
-    useMutation,
-    useQuery,
     warning,
-    withDataProvider,
 };
